tests: userspace: fix read/write_kernel_stack tests
The privileged aka kernel stacks on ARM are not relative to the user
stack, and thus their location cannot be computed directly.  To find
the privileged stack on ARM, we have to use _k_priv_stack_find().
In turn, this requires moving the hash table generated by gperf to
rodata so that we can read it from userspace, so we pass -C to gperf.
Accessing thread_stack directly from the test function requires making
it non-static in ztest, so we also give it a ztest_ prefix to avoid
collisions with other test programs.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>